backend/requirements.txt
absl-py==2.0.0
accelerate==1.9.0
aiofiles==25.1.0
aiohappyeyeballs==2.5.0
aiohttp==3.10.11
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.3
asyncer==0.0.7
attrs==22.2.0
auth0-python==4.10.0
autopy==4.0.1
backcall==0.2.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.11.2
bidict==0.23.1
bitsandbytes==0.45.2
bleach==6.0.0
blinker==1.9.0
blis==0.7.11
build==1.2.2.post1
cachelib==0.13.0
cachetools==5.3.1
catalogue==2.0.10
certifi==2023.5.7
cffi==1.15.1
chainlit==2.3.0
charset-normalizer==3.1.0
chevron==0.14.0
chromadb==1.0.13
click==8.2.1
cloudpathlib==0.16.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.2
confection==0.1.3
contourpy==1.2.0
crewai==0.130.0
cryptography==45.0.4
ctransformers==0.2.27
cycler==0.12.1
cymem==2.0.8
dart==1.0
dataclasses-json==0.6.7
datasets==4.0.0
debugpy==1.6.6
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.7
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
docstring_parser==0.16
durationpy==0.10
EasyProcess==1.1
emoji==2.15.0
entrypoint2==1.1
et_xmlfile==2.0.0
executing==1.2.0
facenet-pytorch==2.5.3
faiss-cpu==1.11.0.post1
fastapi==0.115.11
fastjsonschema==2.16.3
fer==22.5.1
ffmpeg==1.4
ffmpy==0.4.0
filelock==3.12.4
filetype==1.2.0
Flask==3.1.0
Flask-Caching==2.3.1
Flask-SQLAlchemy==3.1.1
flatbuffers==23.5.26
flutter==0.1
fonttools==4.47.2
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.9.2
gast==0.5.4
git-filter-repo==2.47.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.25.1
google-api-python-client==2.173.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.69.1
gradio==4.44.1
gradio_client==1.3.0
greenlet==3.1.1
groq==0.33.0
grpcio==1.71.0
gTTS==2.4.0
h11==0.16.0
h5py==3.10.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.33.5
humanfriendly==10.0
idna==3.4
imageio==2.34.0
imageio-ffmpeg==0.4.9
importlib_metadata==8.5.0
importlib_resources==6.4.5
imutils==0.5.4
iniconfig==2.0.0
instructor==1.8.3
ipykernel==6.29.2
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.18.2
Jinja2==3.1.6
jiter==0.8.2
joblib==1.3.2
json5==0.12.0
json_repair==0.47.1
jsonpatch==1.33
jsonpickle==4.1.1
jsonpointer==2.3
jsonref==1.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.0.3
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
keras==2.14.0
kiwisolver==1.4.5
kubernetes==33.1.0
langchain==1.0.3
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.3
langchain-groq==1.0.0
langchain-huggingface==1.0.1
langchain-openai==1.0.2
langchain-pinecone==0.2.3
langchain-tests==0.3.14
langchain-text-splitters==1.0.0
langcodes==3.3.0
langdetect==1.0.9
langgraph==1.0.2
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.3.45
Lazify==0.4.0
libclang==16.0.6
litellm==1.72.0
literalai==0.1.103
lxml==6.0.2
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.2
marshmallow==3.26.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mediapipe==0.10.5
mistune==2.0.5
ml-dtypes==0.2.0
mmh3==5.1.0
MouseInfo==0.1.3
moviepy==1.0.3
mpmath==1.3.0
mss==9.0.1
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.10
mypy-extensions==1.0.0
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.7.3
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
notebook==6.5.3
notebook_shim==0.2.2
npm==0.1.1
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
onnxruntime==1.22.0
openai==1.109.1
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
openpyxl==3.1.5
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp==1.29.0
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-instrumentation==0.50b0
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opt-einsum==3.3.0
optional-django==0.1.0
orjson==3.10.7
ormsgpack==1.12.0
overrides==7.7.0
packaging==23.2
pandas==2.3.0
pandocfilters==1.5.0
parso==0.8.3
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
peft==0.13.2
pi==0.1.2
pickleshare==0.7.5
pillow==12.0.0
pinecone==5.4.2
pinecone-plugin-inference==3.1.0
pinecone-plugin-interface==0.0.7
pipenv==2024.4.1
platformdirs==4.3.6
plotly==5.19.0
pluggy==1.5.0
posthog==5.3.0
preshed==3.0.9
proglog==0.1.10
prometheus-client==0.16.0
prompt-toolkit==3.0.38
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.5
psutil==5.9.4
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==15.0.2
pyarrow-hotfix==0.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
PyAutoGUI==0.9.54
pybase64==1.4.1
pycparser==2.21
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pydeck==0.8.1b0
pydub==0.25.1
PyGetWindow==0.0.9
Pygments==2.14.0
PyJWT==2.10.1
PyMsgBox==1.0.9
PyMuPDF==1.26.5
pyparsing==3.1.1
pypdf==5.3.1
pypdfium2==4.30.1
pyperclip==1.8.2
PyPika==0.48.9
pyproject_hooks==1.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
pyreadline3==3.5.4
PyRect==0.2.0
pyrsistent==0.19.3
pyscreenshot==3.1
PyScreeze==0.1.29
pytesseract==0.3.13
pytest==8.3.5
pytest-asyncio==0.25.3
pytest-socket==0.7.0
python-dateutil==2.8.2
python-dotenv==1.1.1
python-engineio==4.11.2
python-iso639==2025.2.18
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.18
python-oxmsg==0.0.2
python-socketio==5.12.1
pytweening==1.0.7
pytz==2023.3.post1
pyvis==0.3.2
pywhatkit==5.4
pywin32==305
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.1
qtconsole==5.4.1
QtPy==2.3.0
ragas==0.3.8
RapidFuzz==3.14.3
referencing==0.36.2
regex==2024.11.6
requests==2.32.5
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.25.1
rsa==4.9
ruff==0.6.8
sacremoses==0.1.1
safetensors==0.5.3
scapy==2.4.4
scikit-learn==1.5.2
scikit-network==0.33.3
scipy==1.12.0
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.0
sentence-transformers==5.0.0
sentencepiece==0.2.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
sounddevice==0.5.2
soupsieve==2.4
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.0
SQLAlchemy==2.0.37
srsly==2.4.8
stack-data==0.6.2
starlette==0.41.3
streamlit==1.47.0
sympy==1.12
syncer==2.0.3
syrupy==4.9.0
tavily-python==0.5.4
tenacity==9.1.2
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-cpu==2.14.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
thinc==8.2.1
threadpoolctl==3.2.0
tiktoken==0.9.0
tinycss2==1.2.1
tk==0.1.0
tokenizers==0.21.0
toml==0.10.2
tomli==2.2.1
tomli_w==1.2.0
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.1+cu118
torchaudio==2.2.1+cu118
torchvision==0.17.1
tornado==6.2
tqdm==4.67.1
traitlets==5.9.0
transformers==4.53.3
trl==0.19.1
typer==0.12.5
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2023.3
unstructured==0.18.15
unstructured-client==0.42.3
uptrace==1.29.0
uri-template==1.2.0
uritemplate==4.2.0
urllib3==2.5.0
uv==0.7.13
uvicorn==0.31.0
virtualenv==20.29.3
waitress==3.0.2
wasabi==1.1.2
watchdog==4.0.0
watchfiles==0.20.0
wcwidth==0.2.6
weasel==0.3.3
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
websockets==12.0
Werkzeug==3.1.3
widgetsnbextension==4.0.5
wikipedia==1.4.0
wrapt==1.14.1
wsproto==1.2.0
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
